Fixed a crash bug on AMD64 dealing with integer sizes (bug #895)
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@2410 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/state.h
@@ -32,9 +32,9 @@ extern void InitGroupInfo (BOOLEAN FirstTime);
 extern void UninitGroupInfo (void);
 extern BOOLEAN GetGroupInfo (DWORD offset, BYTE which_group);
 extern DWORD PutGroupInfo (DWORD offset, BYTE which_group);
-#define GROUPS_RANDOM  0L
-#define GROUPS_ADD_NEW (~0L)
-#define GROUP_LIST     0
+#define GROUPS_RANDOM  ((DWORD)(0L))
+#define GROUPS_ADD_NEW ((DWORD)(~0L))
+#define GROUP_LIST     ((BYTE)0)
 #define GROUP_INIT_IP  ((BYTE)~0)
 #define GROUP_LOAD_IP  GROUP_LIST
 #define GROUP_SAVE_IP  ((BYTE)~0)
